[FEAT] 가게 미션 등록 API 구현

+++ spring/umc-spring-study/src/main/java/com/example/umc/repository/MissionRepository.java
@@ -4,6 +4,6 @@ import com.example.umc.domain.Mission;
 import com.example.umc.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MissionRepository extends JpaRepository<Mission, Long> {
+public interface MissionRepository extends JpaRepository<Mission, Integer> {
 }
