+++ spring/umc-spring-study/src/main/java/com/example/umc/repository/StoreRepository.java
@@ -4,6 +4,6 @@ import com.example.umc.domain.Review;
 import com.example.umc.domain.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
 }
